+++ dbt/include/polars/macros/schema_name.sql
@@ -1,6 +1,6 @@
 {% macro polars__generate_schema_name(custom_schema_name, node) -%}
 
-    {%- set default_schema = adapter.get_default_schema(node) -%}
+    {%- set default_schema = target_schema(node=node) -%}
 
     {%- if custom_schema_name is none -%}
 
@@ -12,4 +12,11 @@
 
     {%- endif -%}
 
+{%- endmacro %}
+
+{% macro target_schema(catalog=none, node=none) -%}
+    {%- if node is not none -%}
+        {%- set catalog = (node.config.get("catalog") if node.config else none) or node.database -%}
+    {%- endif -%}
+    {{ return(adapter.get_default_schema({"database": catalog})) }}
 {%- endmacro %}